Code mis au propre dans le dossier Automatisations

+++ Automatisations/nouveau_guerrier.js
@@ -1,9 +1,20 @@
+'use strict';
+
 const sqlite3 = require('sqlite3');
 const db = new sqlite3.Database('database.db')
 
 
 module.exports = {
-    nouveauGuerrier: function (Discord, client, message) {
+    // Increases to 1 the message counter, if the member has more than 300 messages and a difference of 14 days
+    // Between the first message and now, the member is deleted from the table and an embed is sent to bot-meme-spam.
+    nouveauGuerrier: function (client, message) {
+        const embed = {
+            color: 0xffa600,
+            title: 'Félicitations pour ta promotion au rang guerrier <:edt:786348941741654046>',
+            description: `<@${message.author.id}> Tu peux dés à présent avoir accès au salon <#875783142638759976>, aux salons vocaux, mais aussi mettre en ligne les fichiers multimédias de ton choix. \nMerci pour ton activité et ton investissement, tu es un véritable tigre.\n\nQue la masse <:peaky2:678665593490243635>`
+        }
+
+
         db.serialize(function () {
             message.author.username = message.author.username.replace("'", "''")
             db.run(`INSERT OR IGNORE INTO nouveau_guerrier (id, username, count, date) VALUES('${message.author.id}', '${message.author.username}', 0, '${new Date()}');`)
@@ -12,22 +23,14 @@ module.exports = {
                 if (err) {console.log(`ng auto : ${message.author.id}`);console.log(err)}
                 else {
                     row = row[0]
-                    if (row.count >= 300) {
-                        const date = new Date()
-                        const ngdate = new Date(row.date)
-                        if ((Math.abs(date.getTime() - ngdate.getTime()) / (1000 * 3600 * 24)) >= 14) {
+                    const [dateNow, dateJoin] = [new Date(), new Date(row.date)]
+                    const dateDiff = (Math.abs(dateNow.getTime() - dateJoin.getTime()) / (8.64 * 10**7))
+                    if (row.count >= 300 && dateDiff >= 14) {
                             console.log(`${row.username} : ${row.count}`)
                             message.member.roles.add("610959663865004050");
                             message.member.roles.remove("747705073458610197");
                             db.run(`DELETE FROM nouveau_guerrier WHERE id = ${message.author.id};`)
-                            var target_chan = client.channels.resolve('611864040699985931');
-                            target_chan.send('<@' + message.author.id + '>')
-                            const Embed = new Discord.MessageEmbed()
-                                .setColor('#ffa600')
-                                .setTitle('Félicitations pour ta promotion au rang guerrier <:edt:786348941741654046>')
-                                .setDescription('<@' + message.author.id + '> Tu peux dés à présent avoir accès au salon <#875783142638759976>, aux salons vocaux, mais aussi mettre en ligne les fichiers multimédias de ton choix. \nMerci pour ton activité et ton investissement, tu es un véritable tigre.\n\nQue la masse <:peaky2:678665593490243635>')
-                            target_chan.send({ embeds: [Embed] })
-                        }
+                            client.channels.resolve('611864040699985931').send({ content: `<@${message.author.id}>`, embeds: [embed] })
                     }
                 }
             })
